LP1798628: Copy alert counts in Holdings
Currently the count of all copy alerts ever applied
to a copy is displayed in the Holdings view, which
can be quite confusing when clicking Manage and not
seeing the expected number of alerts. This branch
changes the display to only count the number of
events currently active since those are the only
ones available to manage.

Signed-off-by: Jason Boyer <jboyer@library.in.gov>
Signed-off-by: Garry Collum <gcollum@gmail.com>
Signed-off-by: Chris Sharp <csharp@georgialibraries.org>