LP1808268 eg2 grid action disableOnRows sanity check
The === null test for determining if a disableOnRows handler was set was
not broad enough to capture undefined values.  This commit changes it to
a more generic if() test to cover null/undefined.  Also, instead of
creating a dummy function in the action, the testing code checks for the
presence of the function at run time.

Apply a fix to an ng-lint warning introduced by this series of commits in
the sandbox code.

Signed-off-by: Bill Erickson <berickxx@gmail.com>