LP1806087 Ang catalog pending tabs offer manual redirect
Instead of sending users to the AngJS versions of not-yet-implemented
tabs from the Angular catalog, show a message indicating the tab is not
yet implemented and offer a link to open the requested record tab in a
new browser tab.  This better clarifies what is happening and helps to
avoid situations where the back-button is confusingly hijacked via
intermediate redirect.

We now also open the "View" link for copies in a new tab as well for
consistency with the modified "Edit" link.

Signed-off-by: Bill Erickson <berickxx@gmail.com>
Signed-off-by: Dan Wells <dbw2@calvin.edu>