LP#1817332 Add example to lib.timezone setting description
This commit adds a few examples of acceptable lib.timezone values to the
setting description, with the goal of using examples of existing
Evergreen installations. It also includes a clickable link to the
extensive list on Wikipedia. Since the settings editor is displayed
within an iframe, the link needs a "target" to successfully trigger.

Signed-off-by: Remington Steed <rjs7@calvin.edu>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>
Signed-off-by: Bill Erickson <berickxx@gmail.com>