LP#1768869 - Update make_release to add auditor table update
The auditor.update_auditors() function needs to be called
to update tables that are being auditied with schema
changes made to those tables.  This adds a call to that
funcation at the end of the upgrade scripts.

The function seems to not take that long to run, so it may
be ok to run it more than necessary.  Or it can be removed
by release manager if they know that no schema changes were made.

Signed-off-by: Josh Stompro <stompro@stompro.org>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>