LP1721109: Update item status when holdings are edited
To test:
1) Go to Item Status List View, select an item, and go to Actions >
Edit > Items (or Call Numbers, or both).
2) Make a change to one of the columns you have active in
the Item Status list view. Click Save and Exit.
3) Note that the list view of Item Status has not updated.
4) Load this commit.
5) Repeat steps 1-2.
6) Note that the columns are updated this time.  Note that
there is a new status column, and that an icon appears for
all the rows that you have changed.  Also, note that there
is a toast message in the bottom right hand corner of the
screen confirming that you have successfully made your change.

Signed-off-by: Jane Sandberg <sandbej@linnbenton.edu>
Signed-off-by: Chris Sharp <csharp@georgialibraries.org>