LP1716473 Copy editor hide disabled defaults
Disabling a field in the copy editor "Defaults" tab now hides the field
instead of simply disabling it.  Once a field is hidden, subsequent
field shift left to fill the space.

The "Working Items" grid is shifted to the bottom of the page, below the
attributes editor.  This allows the attributes to fill the horizontal
space and puts the working items and save operations visually at the end
of the page, i.e. at end of the work flow.

Signed-off-by: Bill Erickson <berickxx@gmail.com>
Signed-off-by: Geoff Sams <gsams@roanoketexas.com>
Signed-off-by: Michele Morgan <mmorgan@noblenet.org>