LP#1819179: Angular value formatter gets link smarts
Teach the FormatService to display selector values for linked objects
when the requested field is a link field and it contains an object value
instead of just a id/key value.

Improve IDL data passing from the grid to the IDL service so it can
better determine which fields are avialble for link selector display.

Signed-off-by: Bill Erickson <berickxx@gmail.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Dan Wells <dbw2@calvin.edu>