LP1819179 PCRUD selector fleshing handles maps
Teach the PCUD selector fleshing code to handle selector fields on
mapped classes, where an intermediate object flesh is performed by pcrud
at flesh time.

Adds a 'selector' column on metabib.metarecord in the IDL so that we can
have a functioning example of this to use in the sandbox test code.

Signed-off-by: Bill Erickson <berickxx@gmail.com>
Signed-off-by: Dan Wells <dbw2@calvin.edu>