Coerce numbers for bib IDs in Angular staff catalog
Under some circumstances, the search results screen in the new
experimental Angular staff catalog did not display any results (though
it did show the pager, facets, etc.).

This was traced back to the results array attempting to be populated
with ID values of '-1' instead of the real IDs.  Coercing the incoming
IDs to numbers appears to resolve the issue.

Signed-off-by: Bill Erickson <berickxx@gmail.com>
Signed-off-by: Dan Wells <dbw2@calvin.edu>