LP1761142 Volume Copy Editor Changed Fields Hard to Discern
Add a changed_fields object to the scope and update it as copies are
changed. The function field_changed(field) determines if any of the
selected copies have altered the given field.  finally, a CSS class,
field-changed, is applied to add a green border to fields whose values
have been modified.

Signed-off-by: a. bellenir <abelleni@grpl.org>
Signed-off-by: Ruth Frasur <rfrasur@gmail.com>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>