LP1823393 FMEditor component i18n translate buttons
When editing a record in the fieldmapper editor component, offer a
translate option for each i18n field on the object, in the form of a
button next to the field.  Once clicked, the translate component dialog
appears and allows the user to add/modify translations on the selected
field.

Signed-off-by: Bill Erickson <berickxx@gmail.com>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>
Signed-off-by: Jane Sandberg <sandbej@linnbenton.edu>