LP#1823983: repeat_delay check removes event definition id filter
This bug has existed since the repeat_delay feature was added, but likely only
rarely was triggered because most events don't have a repeat_delay, and those
that do don't interact with the same target.  However, as more repeatable event
definitions are created, inappropriate interaction becomes more likely.

In this commit we avoid overwriting the whole join condition clause, which
already contains an event definition id filter and needs to retain it.

Included is an unrelated change that uses the new-ish form of
interval_to_seconds that avoids DST boundary shifting issues by passing a
context DateTime object as the second parameter.

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Galen Charlton <gmc@equinoxOLI.org>