LP#1782014 Fix grid path for Title in patron items out
When the Items Out grid was updated to use display fields, the
displayed title was updated (along with the fleshed data in the
javascript) but the grid "path" was not. This commit fixes the path,
which fixes title sorting. (Title sorting broke because the
simple_record.title was no longer being fleshed in via the JS code, so
the grid was attempting to sort on an empty value.)

Signed-off-by: Remington Steed <rjs7@calvin.edu>
Signed-off-by: Jane Sandberg <sandbej@linnbenton.edu>