Docs: fix uneven lengths of code delimiter blocks
The lengths of the starting and ending sets of hyphens are supposed to be the
same. Asciidoc tools like asciidoctor can be less forgiving than the original
asciidoc processor, with less than hilarious results.

It would probably make sense to try to standardize on something like 60 hyphens
to easily flag this problem for deviations from the norm, but for now here's
the result of running the docs through asciidoctor and manually adjusting the
output.

Signed-off-by: Dan Scott <dan@coffeecode.net>