docs: updating copy tags to item tags
Changed all appropriate references from copy to item.  Included references to a couple updated screenshots (I will momentarily put the files in the media folder).  Proposed changing the title of this page as well, but I understand if that change is rejected - I'm not sure how best to check the existing docs for references to this page, but I will propose a change in root as well.