LP#1828456: use open-ils.auth.login for AuthProxy native login
AuthProxy always uses the username for login purposes -- if you provide
a barcode, it will look up the associated username. But for native login
it uses open-ils.auth.authenticate.init+complete, which redundantly and
incorrectly uses a hard-coded barcode pattern of '^\d' to check whether
the identifier it is given is a username or a barcode (see bug 1827296).
If your username happens to be the same as your barcode, this won't be a
problem. But if your barcode is 123456 and your username is 2legit2quit,
native login fall-back will always fail.

The solution is for AuthProxy to abandon the old init+complete method
and use open-ils.auth.login for native login instead. It will treat the
username as a username instead of applying a bad barcode pattern, thus
allowing patrons whose username begins with a number to login.

Release-note: Fixes issue with numeric usernames when AuthProxy falls
back on the native Evergreen login mechanism.

Signed-off-by: Jeff Davis <jeff.davis@bc.libraries.coop>
Signed-off-by: Mike Rylander <mrylander@gmail.com>