LP#1823367 Tone down org unit row coloring
Some of the text was a little hard to see, and the overall effect of
the many colors, while attractive, was also a little distracting. In
addition, the lighter blue colors were not obviously different from
our standard selected-row-highlight blue.

This commit tries a different direction, using some mild green shades
based on the standard Evergreen green color base.  They are easy to
read and still accomplish the overall goal of bringing out the org
units (a little) and the call numbers (a little more).

Signed-off-by: Dan Wells <dbw2@calvin.edu>
Signed-off-by: Bill Erickson <berickxx@gmail.com>
Signed-off-by: Jane Sandberg <sandbej@linnbenton.edu>