LP#1830089: close transaction and update copy status on payment or adjust to zero
When you make a payment that sets the balance owed to zero, Evergreen
closes the transaction and sets the item status to Lost & Paid (if
appropriate).  Adjust to Zero should do the same thing, but hitherto it
would not update the item status.  This commit refactors some code to
ensure that Evergreen gives the same result whether you make a payment
or adjust to zero.

Signed-off-by: Jeff Davis <jeff.davis@bc.libraries.coop>
Signed-off-by: Terran McCanna <tmccanna@georgialibraries.org>
Signed-off-by: Galen Charlton <gmc@equinoxOLI.org>